Refactoring import
