+++ auth/admin.py
@@ -1,13 +1,8 @@
-from datetime import datetime, timedelta
+
 from django.shortcuts import render_to_response
-from django.http import HttpResponseRedirect
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
-from django.db.models import Count
-
-from import_export import resources
 from import_export.admin import ImportExportModelAdmin
-from import_export import fields
 from communication.utils import VumiSmsApi
 from random import randint
 from auth.forms import SendWelcomeSmsForm
@@ -15,7 +10,6 @@ from django import template
 from communication.utils import get_autologin_link
 from auth.models import Learner, SystemAdministrator, SchoolManager,\
     CourseManager, CourseMentor
-from organisation.models import School
 from forms import SystemAdministratorChangeForm, \
     SystemAdministratorCreationForm, SchoolManagerChangeForm,\
     SchoolManagerCreationForm, CourseManagerChangeForm, \
@@ -23,11 +17,9 @@ from forms import SystemAdministratorChangeForm, \
     CourseMentorCreationForm, LearnerChangeForm, LearnerCreationForm
 import koremutake
 from django.contrib.auth.hashers import make_password
-from django.utils.translation import ugettext_lazy as _
-from organisation.models import Course
 from core.models import ParticipantQuestionAnswer
-from core.models import Participant, Class
-
+from auth.resources import LearnerResource
+from auth.filters import CourseFilter, AirtimeFilter
 
 class SystemAdministratorAdmin(UserAdmin):
     # The forms to add and change user instances
@@ -145,142 +137,6 @@ class CourseMentorAdmin(UserAdmin):
                                                 "password2")}),
         ("Region",                  {"fields": ("country", "area", "course")})
     )
-
-
-class LearnerResource(resources.ModelResource):
-    course = fields.Field(column_name=u'course')
-    completed_questions = fields.Field(column_name=u'completed_questions')
-    percentage_correct = fields.Field(column_name=u'percentage_correct')
-
-    class Meta:
-        model = Learner
-        exclude = (
-            'customuser_ptr', 'password', 'last_login', 'is_superuser',
-            'groups', 'user_permissions', 'is_staff', 'is_active',
-            'date_joined', 'unique_token', 'unique_token_expiry'
-            'welcome_message_sent', 'welcome_message'
-        )
-        export_order = (
-            'id',
-            'username',
-            'first_name',
-            'last_name',
-            'email',
-            'mobile',
-            'school',
-            'country',
-            'area',
-            'city',
-            'optin_sms',
-            'optin_email',
-            'completed_questions',
-            'percentage_correct',
-            'course',
-        )
-
-    def dehydrate_school(self, learner):
-        if learner.school is not None:
-            return learner.school.name
-        else:
-            return ""
-
-    def dehydrate_completed_questions(self, learner):
-        return ParticipantQuestionAnswer.objects.filter(
-            participant__learner=learner
-        ).count()
-
-    def dehydrate_percentage_correct(self, learner):
-        complete = self.dehydrate_completed_questions(learner)
-        if complete > 0:
-            return ParticipantQuestionAnswer.objects.filter(
-                participant__learner=learner,
-                correct=True
-            ).count()*100/complete
-        else:
-            return 0
-
-    def get_or_init_instance(self, instance_loader, row):
-        row[u'is_staff'] = False
-        row[u'is_superuser'] = False
-        row[u'is_active'] = True
-        row[u'date_joined'] = datetime.now()
-        row[u'welcome_message_sent'] = None
-        row[u'welcome_message'] = None
-
-        return super(resources.ModelResource, self) \
-            .get_or_init_instance(instance_loader, row)
-
-    def import_obj(self, obj, data, dry_run):
-        school, created = School.objects.get_or_create(name=data[u'school'])
-        data[u'school'] = school.id
-        return super(resources.ModelResource, self)\
-            .import_obj(obj, data, dry_run)
-
-    def save_m2m(self, obj, data, dry_run):
-        classs = Class.objects.filter(name=data[u'course']).first()
-
-        # If the course and respective class exist, create participant
-        if classs and not dry_run:
-            Participant.objects.create(
-                learner=obj,
-                classs=classs,
-                datejoined=datetime.now(),
-
-            )
-
-        return super(resources.ModelResource, self)\
-            .save_m2m(obj, data, dry_run)
-
-
-class CourseFilter(admin.SimpleListFilter):
-    title = _('Course')
-    parameter_name = 'id'
-
-    def lookups(self, request, model_admin):
-        return Course.objects.all().values_list('id', 'name')
-
-    def queryset(self, request, queryset):
-        if self.value() is None:
-            return queryset
-        else:
-            return queryset.filter(participant__classs__course_id=self.value())
-
-
-class AirtimeFilter(admin.SimpleListFilter):
-    title = _('Airtime')
-    parameter_name = 'name'
-
-    def get_date_range(self):
-        today = datetime.today()
-        start = today - timedelta(days=today.weekday(), weeks=1)
-        end = start + timedelta(days=7)
-        return [start, end]
-
-    def get_learner_ids(self):
-        participants = ParticipantQuestionAnswer.objects.filter(
-            answerdate__range=self.get_date_range()
-        ).values('participant').annotate(Count('participant'))
-
-        filtered_participants = [
-            i for i in participants
-            if i['participant__count'] >= 9]
-
-        return Participant.objects.filter(
-            id__in=filtered_participants
-        ).values_list('learner', flat=True)
-
-    def lookups(self, request, model_admin):
-        return [('airtime_award', _('9 to 15 questions correct'))]
-
-    def queryset(self, request, queryset):
-        if self.value() is None:
-            return queryset
-        else:
-            if self.value() == 'airtime_award':
-                learners = self.get_learner_ids()
-                return queryset.filter(id__in=learners)
-            else:
-                return queryset
 
 
 class LearnerAdmin(UserAdmin, ImportExportModelAdmin):
